Missing cat being passed to the function?
I am working through these examples as an intro to this js lib and when running the function gameLoop, firefox's javascript console complained about cat not being defined.
I could be wrong here, it is a long time since I was introduced to javascript but I think it has something to do with scope and calling a function within a function?
Please have a look and see what you think. Maybe I missed something?
Thanks.